feat: extract eip1193 retrieval logic
